% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naiveRCIPerm.R
\name{naiveRCIPerm}
\alias{naiveRCIPerm}
\title{Naive RCI Permutations}
\usage{
naiveRCIPerm(x, y, delta_x = 0.2, delta_y = 0.2,
  valid.logic = c("and", "or"), tie.method.x = c("ignore", "half"),
  tie.method.y = c("ignore", "half"), required_alpha = 1e-06,
  p_confidence = 0.01, C = TRUE)
}
\description{
Naive RCI Permutations
}
