% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paired.concordance.index.R
\name{paired.concordance.index}
\alias{paired.concordance.index}
\title{Takes two numerical vectors and computes the concordance index between them
by comparing the order of values for two pairs of data each time}
\usage{
paired.concordance.index(
  predictions,
  observations,
  delta.pred = 0,
  delta.obs = 0,
  alpha = 0.05,
  outx = FALSE,
  alternative = c("two.sided", "less", "greater"),
  logic.operator = c("and", "or"),
  CPP = TRUE,
  p_method = c("Permutation", "Asymptotic"),
  conf_int_method = c("Bootstrap", "Asymptotic"),
  num_hypothesis = 1,
  perm_p_confidence = 0.2,
  boot_num = 5000,
  comppairs = 10
)
}
\arguments{
\item{predictions}{{numeric} A vector of predicted drug responces which could
be either continuous or discrete}

\item{observations}{{numeric} A vector of observed continuous drug responces}

\item{delta.pred}{{numeric} The minimunm reliable difference between two
values in the predictions vector to be considered as significantly various
values.}

\item{delta.obs}{{numeric} The minimunm reliable difference between two
values in the observations vector to be considered as significantly various
values. In drug sensitivity , default value for delta.pred is picked by
looking into delta auc values (drug response metric) between biological
replicates across three large pharmacogenomic studies,
CTRPv2 (370 drugs over ~15-20 cells), GDSC (1 drug over ~600 cells),
GRAY (85 drugs over ~10-50 cells)}

\item{alpha}{{numeric} alpha level to compute confidence interval}

\item{outx}{{boolean} set to TRUE to not count pairs of predictions that are
tied as a relevant pair. This results in a Goodman-Kruskal gamma type rank
correlation.}

\item{alternative}{{character} What is the alternative hypothesis? Must be
one of "two.sides", "less", and "greater" and defaults to two.sides".}

\item{logic.operator}{{character} determines how strict should the test be to
remove noisy pairs. Must be one of "and" or "or" and defaults to "and".}

\item{CPP}{{boolean} Whether to use the C version of the code for faster
execution}

\item{p_method}{{character} Either "Permutation", or "Asymptotic", picks a
method to use for calculating p-values. If Permutation, then "alpha"/"num_hypothesis"
is used to determine the effective alpha used for estimating number of required 
permutations.}

\item{conf_int_method}{{character} Either "Bootstrap" or "Asymptotic", picks a method for
estimating the confidence interval corresponding to 1-"alpha".}

\item{num_hypothesis}{{numeric} Total number of hypothesis being tested in analysis. Used
for adjusting number of required permutations when using the permutation method of computing
p values. Default 1. Ignored if using asymptotic p value.}

\item{perm_p_confidence}{{numeric} Maximum permited 1 SD confidence interval of our estimated 
permutation p value around the true p value, as a fraction of "alpha"/"num_hypothesis". Ignored
if using asymptotic p value, no guarantee on correctness exists.}

\item{boot_num}{{numeric} number of samples to use for bootstrap. Default 5000. Ignored
if using asymptotic confidence interval.}

\item{comppairs}{{numeric} minimum number of pairs to calculate a valid CI.}
}
\value{
[list] ! list of concordance index and its pvalue
along with the lower and upper confidence intervals
}
\description{
This function returns the concordance index and its p-value
along with the lower and upper confidence intervals of said p-value.
}
\examples{
data(PLX4720_data)
pci_PLX4720 <- paired.concordance.index(predictions = PLX4720_data[ ,"AAC_CTRPv2"],
observations = PLX4720_data[ ,"AAC_GDSC"], delta.pred = 0, delta.obs = 0,
outx = TRUE)
pci_PLX4720$cindex

}
