% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLX4720_data.R
\docType{data}
\name{PLX4720_data}
\alias{PLX4720_data}
\title{PLX4720 Dataset}
\format{A dataframe with 534 rows and 2 variables}
\usage{
PLX4720_data
}
\description{
A dataset for use in wCI package examples
}
\keyword{datasets}
