% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naiveRCIPerm.R
\name{naiveRCIPerm}
\alias{naiveRCIPerm}
\title{Naive RCI Permutations}
\usage{
naiveRCIPerm(x, y, delta_x = 0.2, delta_y = 0.2,
  tie.method.x = c("ignore", "half"), tie.method.y = c("ignore",
  "half"), alternative = c("two.sided", "greater", "less"),
  required_alpha = 1e-06, p_confidence = 0.01, C = TRUE,
  verbose = FALSE)
}
\description{
Naive RCI Permutations
}
