% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paired.concordance.index.R
\name{paired.concordance.index}
\alias{paired.concordance.index}
\title{Takes two numerical vectors and computes the concordance index between them
by comparing the order of values for two pairs of data each time}
\usage{
paired.concordance.index(predictions, observations, delta.pred = 0,
  delta.obs = 0, alpha = 0.05, outx = FALSE,
  alternative = c("two.sided", "less", "greater"),
  logic.operator = c("and", "or"), CPP = TRUE, comppairs = 10)
}
\arguments{
\item{predictions}{{numeric} A vector of predicted drug responces which could be either continuous or discrete}

\item{observations}{{numeric} A vector of observed continuous drug responces}

\item{delta.pred}{{numeric} The minimunm reliable difference between two values in the predictions vector to be considered as significantly various values.}

\item{delta.obs}{{numeric} The minimunm reliable difference between two values in the observations vector to be considered as significantly various values.
In drug sensitivity , default value for delta.pred is picked by looking into delta auc values (drug response metric) between biological replicates across three
large pharmacogenomic studies, CTRPv2(370 drugs over ~15-20 cells) , GDSC(1 drug over ~600 cells), GRAY (85 drugs over ~10-50 cells)}

\item{alpha}{{numeric} alpha level to compute confidence interval}

\item{outx}{{boolean} set to TRUE to not count pairs of predictions that are tied as a relevant pair.
This results in a Goodman-Kruskal gamma type rank correlation.}

\item{alternative}{{character} what is the alternative hypothesis? Must be one of "two.sides", "less", and "greater" and defaults to two.sides".}

\item{logic.operator}{{character} determines how strict should the test be to remove noisy pairs. Must be one of "and" or "or" and defaults to "and".}

\item{CPP}{{boolean} whether to use the C version of the code for faster execution}

\item{comppairs}{{numeric} minimum number of pairs to calculate a valid CI}
}
\value{
[list] ! list of concordance index and its pvalue
along with the lower and upper confidence intervals
}
\description{
This function return the concordance index and its p-value
along with the lower and upper confidence intervals of said p-value.
}
\examples{
gdsc.auc <- PharmacoGx::summarizeSensitivityProfiles(GDSC, sensitivity.measure='auc_published')
xx <- PharmacoGx::summarizeSensitivityProfiles(predicted.by.model, gdsc.auc$Erlotinib)

}
