% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paired.concordance.index.R
\name{paired.concordance.index}
\alias{paired.concordance.index}
\title{Takes two numerical vectors and computes the concordance index between them
by comparing the order of values for two pairs of data each time}
\usage{
paired.concordance.index(predictions, observations, delta.pred = 0.2,
  delta.obs = 0.2, alpha = 0.05, outx = TRUE,
  alternative = c("two.sided", "less", "greater"), logic.operator = c("and",
  "or"))
}
\arguments{
\item{predictions}{{numeric} A vector of predicted drug responces which could be either continuous or discrete}

\item{observations}{{numeric} A vector of observed continuous drug responces}

\item{alpha}{{numeric} alpha level to compute confidence interval}

\item{outx}{{boolean} set to TRUE to not count pairs of observations tied on x as a relevant pair.
This results in a Goodman-Kruskal gamma type rank correlation.}

\item{alternative}{{character} what is the alternative hypothesis? Must be one of "two.sides", "less", and "greater".}

\item{cutoff}{{numeric} A drug responce threshold which is used to classify cells to sensitive vs resistant to drug.}

\item{delta}{{numeric} The minimunm reliable difference between two drug sensitivity values to be considered as significantly various responses.
default value for delta is picked by looking into delta auc values between biological replicates across three
large pharmacogenomic studies, CTRPv2(370 drugs over ~15-20 cells) , GDSC(1 drug over ~600 cells), GRAY (85 drugs over ~10-50)}
}
\value{
[list] ! list of concordance index and its pvalue
along with the lower and upper confidence intervals
}
\description{
This function return the concordance index and its p-value
along with the lower and upper confidence intervals of said p-value.
}
\examples{
gdsc.auc <- summarizeSensitivityProfiles(GDSC, sensitivity.measure=‘auc_published’)
xx <- summarizeSensitivityProfiles(predicted.by.model, gdsc.auc$Erlotinib)

}
